refactor: prepare for RuleSet interface
Moves RuleSet into a RuleSet subnamespace.
